メッセージ送信処理を切り出した
