+++ frontend/src/features/chat/utils.ts
@@ -1,19 +1,38 @@
-import { fetchWithToken } from '@/lib/axios'
-import { Room } from '@/types'
-import Error from 'next/error'
-import useSWR from 'swr'
+import { apiClient } from '@/lib/axios'
+import { User } from '@/lib/jotai/userState'
+import { sendMessage } from '@/utils/socket'
 
-export const useRoom = (
+type HandleSendMessageProps = {
+  inputMessage: string
   roomId: string
-): {
-  room: Room
-  error: Error
-  isLoading: boolean
-} => {
-  const { data, error, isLoading } = useSWR(`/rooms/${roomId}`, fetchWithToken)
-  return {
-    room: data?.data.room,
-    error,
-    isLoading,
+  currentUser: User | null
+  setInputMessage: (input: string) => void
+}
+
+export const handleSendMessage = async ({
+  inputMessage,
+  roomId,
+  currentUser,
+  setInputMessage,
+}: HandleSendMessageProps) => {
+  if (inputMessage.trim() === '' || currentUser?.id === undefined) return
+
+  const createMessageArgs = {
+    body: inputMessage,
+    roomId: roomId,
+    userId: currentUser.id,
   }
+  try {
+    const res = await apiClient.post('/messages', createMessageArgs)
+    sendMessage({
+      message: {
+        ...res.data.message,
+        roomId: roomId,
+        userId: currentUser.id,
+      },
+    })
+  } catch (error) {
+    console.error('Failed to send message:', error)
+  }
+  setInputMessage('')
 }
